app/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn==21.2.0
boto3==1.35.0
